eslint and fixes
