+++ package.json
@@ -9,11 +9,12 @@
     "npm": "^10.2.3"
   },
   "scripts": {
-    "start": "vite",
-    "build": "vite build",
+    "start": "npm run getEvents && vite",
+    "build": "npm run lint && vite build",
     "preview": "vite preview",
-    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-    "getEvents": "node ./src/googleapi2events.js"
+    "lint": "eslint ./src --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "lint-fix": "eslint ./src --ext js,jsx --report-unused-disable-directives --max-warnings 0 --fix",
+    "getEvents": "node ./scripts/googleapi2events.js"
   },
   "dependencies": {
     "@fullcalendar/core": "^6.1.10",
@@ -21,11 +22,11 @@
     "@fullcalendar/icalendar": "^6.1.10",
     "@fullcalendar/interaction": "^6.1.10",
     "@fullcalendar/react": "^6.1.10",
+    "@fullcalendar/rrule": "^6.1.10",
     "@fullcalendar/timegrid": "^6.1.10",
     "@mdi/js": "^7.3.67",
     "@mdi/react": "^1.6.1",
     "googleapis": "^131.0.0",
-    "@fullcalendar/rrule": "^6.1.10",
     "html-react-parser": "^5.0.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
@@ -35,11 +36,13 @@
     "@types/react-dom": "^18.2.17",
     "@vitejs/plugin-react": "^4.0.3",
     "eslint": "^8.45.0",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
-    "vite": "^5.0.5",
+    "html-to-text": "^9.0.5",
     "ical-generator": "^6.0.0",
-    "html-to-text": "^9.0.5"
+    "sort-imports": "^1.1.0",
+    "vite": "^5.0.5"
   }
 }
